Abstract plugin directory function into own module and use in versioning.py
