+++ app/resonant-laboratory/server/pluginDir.py
@@ -1,0 +1,10 @@
+import inspect
+import os
+
+
+def pluginDir():
+    me = inspect.stack()[0][1]
+    mydir = os.path.dirname(me)
+    pluginDir = os.path.join(mydir, '..')
+
+    return os.path.abspath(pluginDir)
